rerender sidebar
